+++ tests/testthat/test_Utilities.R
@@ -1,9 +1,5 @@
 # This file contains tests for functions in Utilities.R.
 
-###########################################################
-context("Organizing arguments")
-###########################################################
-
 test_that("errors are generated when organize_args is called with baloney", {
   expect_error(matsbyname:::organize_args(b = 42), 
                "Missing argument a with no fill in organize_args.")
@@ -30,10 +26,6 @@ test_that("oddball match_type works as expected", {
                list(a = matrix(1), b = matrix(2)))
 })
 
-
-###########################################################
-context("Selecting rows and columns")
-###########################################################
 
 test_that("an error is generated when no retain or remove patterns are default", {
   # Check with non-NULL values for a.
@@ -76,6 +68,7 @@ test_that("selecting rows and columns works even when there is a NULL situation"
   expect_null(m %>% select_cols_byname(retain_pattern = "c3"))
 })
 
+
 test_that("setting row and column names works even when there is a NULL situation", {
   m <- matrix(1:4, nrow = 2, ncol = 2, dimnames = list(c("r1", "r2"), c("c1", "c2"))) %>% 
     setrowtype("rows") %>% setcoltype("cols")
@@ -91,10 +84,6 @@ test_that("setting row and column names works even when there is a NULL situatio
   expect_null(setcolnames_byname(NULL, c("a", "b")))
 })
   
-
-###########################################################
-context("Cleaning")
-###########################################################
 
 test_that("bad margins in clean_byname work as expected", {
   m <- matrix(c(0, 0, 0, 1, 2, 3), nrow = 3, ncol = 2, dimnames = list(c("r1", "r2", "r3"), c("c1", "c2")))
@@ -155,10 +144,6 @@ test_that("cleaning works with tolerance", {
 })
 
 
-###########################################################
-context("Is zero")
-###########################################################
-
 test_that("iszero_byname works as expected", {
   m <- matrix(0, nrow = 3, ncol = 2)
   expect_true(iszero_byname(m))
@@ -166,11 +151,6 @@ test_that("iszero_byname works as expected", {
   expect_false(iszero_byname(n))
 })
 
-
-
-###########################################################
-context("Row and column naming")
-###########################################################
 
 test_that("getting row names works as expected", {
   m <- matrix(c(1:6), nrow = 2, dimnames = list(paste0("i", 1:2), paste0("p", 1:3))) %>%
@@ -348,16 +328,16 @@ test_that("renaming rows to prefix or suffix works as expected", {
               dimnames = list(c("a -> b", "r2", "r3"), c("c1", "c2")))
   expected <- m
   rownames(expected) <- c("a", "r2", "r3")
-  actual <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 1, notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 1, notation = RCLabels::arrow_notation)
   expect_equal(actual, expected)
   
   expected <- m
   rownames(expected) <- c("b", "r2", "r3")
-  actual <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 1, notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 1, notation = RCLabels::arrow_notation)
   expect_equal(actual, expected)
   
   # Check that renaming works for a list
-  actual <- rename_to_pref_suff_byname(list(m, m), keep = "suffix", margin = 1, notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(list(m, m), keep = "suffix", margin = 1, notation = RCLabels::arrow_notation)
   expect_equal(actual, list(expected, expected))
 })
 
@@ -369,21 +349,21 @@ test_that("renaming columns to prefix or suffix works as expected", {
               dimnames = list(c("a -> b", "r2", "r3"), c("a -> b", "c -> d")))
   expected <- m
   colnames(expected) <- c("a", "c")
-  actual <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 2, notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 2, notation = RCLabels::arrow_notation)
   expect_equal(actual, expected)
   
   expected <- m
   colnames(expected) <- c("b", "d")
-  actual <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 2, notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 2, notation = RCLabels::arrow_notation)
   expect_equal(actual, expected)
   
   # Check that renaming works for a list
-  actual <- rename_to_pref_suff_byname(list(m, m), keep = "suffix", margin = 2, notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(list(m, m), keep = "suffix", margin = 2, notation = RCLabels::arrow_notation)
   expect_equal(actual, list(expected, expected))
   
   # Check that row and column types are preserved
   m <- m %>% setrowtype("Rows -> Cols") %>% setcoltype("Cols -> Rows")
-  res <- rename_to_pref_suff_byname(m, keep = "suffix", notation = arrow_notation())
+  res <- rename_to_pref_suff_byname(m, keep = "suffix", notation = RCLabels::arrow_notation)
   expect_equal(rowtype(res), "Cols")
   expect_equal(coltype(res), "Rows")
 })
@@ -398,22 +378,22 @@ test_that("renaming rows and columns to prefix or suffix works as expected", {
   rownames(expected) <- c("a", "r2", "r3")
   colnames(expected) <- c("a", "c")
   # Default is margin = c(1, 2)
-  actual <- rename_to_pref_suff_byname(m, keep = "prefix", notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "prefix", notation = RCLabels::arrow_notation)
   expect_equal(actual, expected)
   
   expected <- m
   rownames(expected) <- c("b", "r2", "r3")
   colnames(expected) <- c("b", "d")
-  actual <- rename_to_pref_suff_byname(m, keep = "suffix", notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "suffix", notation = RCLabels::arrow_notation)
   expect_equal(actual, expected)
   
   # Check that renaming works for a list
-  actual <- rename_to_pref_suff_byname(list(m, m), margin = list(c(1, 2)), keep = "suffix", notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(list(m, m), margin = list(c(1, 2)), keep = "suffix", notation = RCLabels::arrow_notation)
   expect_equal(actual, list(expected, expected))
 
   # Check that row and column types are preserved
   m <- m %>% setrowtype("Rows") %>% setcoltype("Cols")
-  res <- rename_to_pref_suff_byname(m, keep = "prefix", notation = arrow_notation())
+  res <- rename_to_pref_suff_byname(m, keep = "prefix", notation = RCLabels::arrow_notation)
   expect_equal(rowtype(res), "Rows")
   expect_equal(coltype(res), "Cols")
 })
@@ -425,13 +405,13 @@ test_that("renaming rows and cols to pref and suff also changes rowtype and colt
                 5, 6), nrow = 3, byrow = TRUE, 
               dimnames = list(c("a -> b", "c -> d", "e -> f"), c("g -> h", "i -> j"))) %>% 
     setrowtype("Industry -> Product") %>% setcoltype("Product -> Industry")
-  res <- rename_to_pref_suff_byname(m, keep = "prefix", notation = arrow_notation())
+  res <- rename_to_pref_suff_byname(m, keep = "prefix", notation = RCLabels::arrow_notation)
   expect_equal(rownames(res), c("a", "c", "e"))
   expect_equal(colnames(res), c("g", "i"))
   expect_equal(rowtype(res), "Industry")
   expect_equal(coltype(res), "Product")
 
-  res2 <- rename_to_pref_suff_byname(m, keep = "suffix", notation = arrow_notation())
+  res2 <- rename_to_pref_suff_byname(m, keep = "suffix", notation = RCLabels::arrow_notation)
   expect_equal(rownames(res2), c("b", "d", "f"))
   expect_equal(colnames(res2), c("h", "j"))
   expect_equal(rowtype(res2), "Product")
@@ -445,11 +425,11 @@ test_that("changing row and column type correctly ignores missing suffixes", {
                 5, 6), nrow = 3, byrow = TRUE, 
               dimnames = list(c("a -> b", "c -> d", "e -> f"), c("g -> h", "i -> j"))) %>% 
     setrowtype("Rows") %>% setcoltype("Product -> Industry")
-  res <- rename_to_pref_suff_byname(m, keep = "prefix", notation = arrow_notation())
+  res <- rename_to_pref_suff_byname(m, keep = "prefix", notation = RCLabels::arrow_notation)
   expect_equal(rowtype(res), "Rows")
   expect_equal(coltype(res), "Product")
 
-  res_2 <- rename_to_pref_suff_byname(m, keep = "suffix", notation = arrow_notation())
+  res_2 <- rename_to_pref_suff_byname(m, keep = "suffix", notation = RCLabels::arrow_notation)
   expect_equal(rowtype(res_2), "Rows")
   expect_equal(coltype(res_2), "Industry")
 })
@@ -464,7 +444,7 @@ test_that("setting identical row/col names is OK", {
   rownames(expected) <- c("a", "a", "r3")
   colnames(expected) <- c("a", "a")
   # The next call will create duplicate row names and column names in m. 
-  actual <- rename_to_pref_suff_byname(m, keep = "prefix", notation = arrow_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "prefix", notation = RCLabels::arrow_notation)
   # Interestingly the command View(actual) or View(expected)
   # shows row names that aren't true. 
   # The 2nd row of actual and expected is shown as "a.1", 
@@ -483,43 +463,43 @@ test_that("renaming with full prefix identifiers works as expected.", {
   expected <- m
   dimnames(expected) <- list(c("a", "c", "e"), c("g", "i"))
   # The next call will create duplicate row names and column names in m. 
-  actual <- rename_to_pref_suff_byname(m, keep = "prefix", notation = bracket_notation())
+  actual <- rename_to_pref_suff_byname(m, keep = "prefix", notation = RCLabels::bracket_notation)
   expect_equal(actual, expected)
   
   expected2 <- m
   dimnames(expected2) <- list(c("a", "c", "e"), c("g [h]", "i [j]"))
-  actual2 <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 1, notation = bracket_notation())
+  actual2 <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 1, notation = RCLabels::bracket_notation)
   expect_equal(actual2, expected2)
   
   expected3 <- m
   dimnames(expected3) <- list(c("a [b]", "c [d]", "e [f]"), c("g", "i"))
-  actual3 <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 2, notation = bracket_notation())
+  actual3 <- rename_to_pref_suff_byname(m, keep = "prefix", margin = 2, notation = RCLabels::bracket_notation)
   expect_equal(actual3, expected3)
   
   expected4 <- m
   dimnames(expected4) <- list(c("b", "d", "f"), c("h", "j"))
-  actual4 <- rename_to_pref_suff_byname(m, keep = "suffix", notation = bracket_notation())
+  actual4 <- rename_to_pref_suff_byname(m, keep = "suffix", notation = RCLabels::bracket_notation)
   expect_equal(actual4, expected4)
 
   expected5 <- m
   dimnames(expected5) <- list(c("b", "d", "f"), c("g [h]", "i [j]"))
-  actual5 <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 1, notation = bracket_notation())
+  actual5 <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 1, notation = RCLabels::bracket_notation)
   expect_equal(expected5, actual5)
 
   expected6 <- m
   dimnames(expected6) <- list(c("a [b]", "c [d]", "e [f]"), c("h", "j"))
-  actual6 <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 2, notation = bracket_notation())
+  actual6 <- rename_to_pref_suff_byname(m, keep = "suffix", margin = 2, notation = RCLabels::bracket_notation)
   expect_equal(expected6, actual6)
   
   # Try with a list
-  actual_list <- rename_to_pref_suff_byname(list(m, m), keep = "prefix", margin = 1, notation = bracket_notation())
+  actual_list <- rename_to_pref_suff_byname(list(m, m), keep = "prefix", margin = 1, notation = RCLabels::bracket_notation)
   expect_equal(actual_list[[1]], expected2)
   expect_equal(actual_list[[2]], expected2)
   
   # Try in a data frame
   DF <- tibble::tibble(m = list(m, m, m, m, m, m), keep = c("prefix", "prefix", "prefix", "suffix", "suffix", "suffix"), 
                        margin = list(c(1, 2), 1, 2, c(1, 2), 1, 2),
-                       notation = list(bracket_notation()),
+                       notation = list(RCLabels::bracket_notation),
                        expected = list(expected, expected2, expected3, expected4, expected5, expected6))
   
   res <- DF %>% 
@@ -530,10 +510,6 @@ test_that("renaming with full prefix identifiers works as expected.", {
   expect_equal(res$actual, res$expected)
 })
 
-
-###########################################################
-context("Row and column types")
-###########################################################
 
 test_that("setrowtype and rowtype works as expected", {
   productnames <- c("p1", "p2")